+++ package.json
@@ -74,7 +74,7 @@
   "scripts": {
     "start": "node scripts/start.js",
     "build": "node scripts/build.js",
-    "test": "xo && node scripts/test.js"
+    "test": "node scripts/test.js"
   },
   "eslintConfig": {
     "extends": "react-app"
@@ -131,10 +131,6 @@
     ]
   },
   "devDependencies": {
-    "eslint-config-xo-react": "^0.18.0",
-    "xo": "^0.24.0"
-  },
-  "xo": {
-    "extends": "xo-react"
+    "eslint-config-xo-react": "^0.18.0"
   }
 }
