+++ src/app/components/Bar/styles/index.tsx
