+++ src/app/components/Sidebar/components/FursonaPicture/index.tsx
