+++ src/app/components/Sidebar/components/FursonaPicture/styles/index.tsx
