+++ src/app/pages/About/styles/index.tsx
@@ -1,0 +1,16 @@
+import { css } from '@emotion/core';
+
+export const Heading = css`
+    color: #fff;
+    padding-left: 30px;
+    font-family: 'Roboto', sans-serif;
+    font-size: 36px;
+    margin-top: 10px;
+`;
+
+export const Subheading = css`
+    color: #fff;
+    padding-left: 30px;
+    margin-top: -10px;
+    font-family: 'Roboto', sans-serif;
+`;
