+++ src/app/pages/Projects/styles/index.tsx
@@ -1,0 +1,12 @@
+import { css } from '@emotion/core';
+
+export const Heading = css`
+    color: #fff;
+    font-family: 'Roboto', sans-serif;
+    font-size: 36px;
+`;
+
+export const Subtitle = css`
+    color: #fff;
+    font-family: 'Roboto', sans-serif;
+`;
